Move global type stuff into a .d.ts file
