+++ pages/graph.tsx
@@ -1,10 +1,10 @@
 import { LoadingIcon } from '../components/LoadingIcon';
 import { Metadata } from '../components/Metadata/Metadata';
 import { MovingAverageOption } from '../types/math';
-import { getCurrentDateString, getDateStringRange } from '../utils/dates';
 import { useAuthenticatedRoute } from '../utils/hooks/firebase';
 import { usePixels } from '../utils/hooks/pixels';
 import { getMovingAverage, MOVING_AVERAGE_OPTIONS } from '../utils/math';
+import { getCurrentDateString, getDateStringRange } from '../utils/time';
 import { ThemeContext } from './_app';
 import {
 	CategoryScale,
