+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-	"include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+	"include": ["next-env.d.ts", "global.d.ts", "**/*.ts", "**/*.tsx"],
 	"exclude": ["node_modules"],
 	"compilerOptions": {
 		"esModuleInterop": true,
