+++ worker/index.ts
@@ -1,5 +1,3 @@
-import { WorkboxGlobalScope } from '../types/environment';
-
-declare let self: WorkboxGlobalScope;
-
 self.__WB_DISABLE_DEV_LOGS = true;
+
+export {};
